Add files via upload
A program that calculates a student's grade point average over their entire course. The student's name is entered on the first line, and his annual grades are on each subsequent line. The student advances to the next grade if his annual grade is greater than or equal to 4.00. If the student is interrupted more than once, the student is expelled and the program ends, printing the name of the student and the class in which he was expelled. A program that calculates a student's grade point average over their entire course. The student's name is entered on the first line, and his annual grades are on each subsequent line. The student advances to the next grade if his annual grade is greater than or equal to 4.00. If the student is interrupted more than once, the student is expelled and the program ends, printing the name of the student and the class in which he was expelled.